Refactor GZipAccessNative to use base class
This reduces the mostly duplicate code that would be used in the SharpZipLib implementation.
